Moved addCoffee to adminDashboard. Fixed duplicate profile link in mobile nav. Moved coffee api call from app.js to Coffees component to avoid needing to refresh page.
