+++ src/components/Auth/Dashboard/AdminDashboard.css
@@ -1,12 +1,16 @@
 .dashboardContent {
     width: 100%;
+    display: flex;
+    margin-top: 45px;
+    justify-content: space-between;
 }
 
 .dashboard-content-container {
-    width: 100%;
+    width: 48%;
     display: flex;
     align-items: center;
     margin-top: 25px;
+    box-shadow: 0 4px 8px 0 rgb(0 0 0 / 20%);
 }
 
 .dashboard-content-container h2 {
@@ -14,13 +18,13 @@
 }
 
 .dashboard-content-coffee-list {
-    margin: 0px 25px;
+    width: 100%;
 }
 
 .dashboard-content-content {
     display: flex;
     justify-content: space-around;
-    margin: 25px 0px;
+    margin: 25px 25px;
 }
 
 .dashboard-content-content p {
@@ -85,4 +89,13 @@
 .back-container:hover, .back-container i:hover {
     color: #0d6efd!important;
     border-bottom: 1px solid #0d6efd;
+}
+
+@media only screen and (max-width: 719px) {
+    .dashboardContent {
+        flex-direction: column;
+    }
+    .dashboard-content-container {
+        width: 100%;
+    }
 }
