+++ src/components/Auth/Dashboard/Dashboard.css
@@ -3,5 +3,5 @@
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    margin: 10%;
+    margin: 5%;
 }
