+++ src/components/Nav/Nav.js
@@ -37,8 +37,8 @@ const Nav = () => {
              </ul>
             {isOpen ? <div className='hamburger-menu hamburger-container'>
                 <ul>
-                    {currentUser ? <Link to='/dashboard'><li>Profile</li></Link> : <Link to='/login'><li>Login</li></Link>}
-                    <Link to='/dashboard'><li onClick={closeMobile}>Profile</li></Link>
+                    {currentUser ? <Link to='/dashboard'><li onClick={closeMobile}>Profile</li></Link> : <Link to='/login'><li onClick={closeMobile}>Login</li></Link>}
+                    
                     <Link to='/'><li onClick={closeMobile}>link1</li></Link>
                     <Link to='/cart'><li onClick={closeMobile}><i className="fas fa-shopping-cart"></i>(1)</li></Link>
                 </ul>
